# dimos/server/openapi.json
{
  "components": {
    "schemas": {
...
      "BlueprintList": {
        "additionalProperties": true,
        "properties": {
          "blueprints": {
            "description": "Built-in blueprints (sorted, without demo-*), then external ones",
            "items": {
              "$ref": "#/components/schemas/BlueprintName"
            },
            "title": "Blueprints",
            "type": "array"
          }
        },
        "required": [
          "blueprints"
        ],
        "title": "BlueprintList",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": true,
        "description": "Every error answer (4xx and 5xx).",
        "properties": {
          "error": {
            "description": "What went wrong, in words a person can act on",
            "examples": [
              "no upload u7"
            ],
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The dimos server's HTTP API: blueprints, global config, runs and their logs, events, Dimensional cloud uploads,\ndiscovery (blueprints, modules, message types), docs, extras and jobs.\n\ndimOS Desktop starts the server (`python -m dimos.server`, on a unix socket) and forwards `/dimos/...` to it unchanged.\nEvery answer is JSON unless the operation says otherwise; every error is `{\"error\": \"<message>\"}` (ErrorResponse).\nState is HTTP; changes are events, published on zenoh at `<ns>/dimos/events/<type>` (the `events` tag).\n\nExtensions: `x-family: dimos` on every operation; `x-agent: true` marks what Desktop's agent finds through\nsearch_endpoints; `x-mcp-tool` names an MCP tool that does the same; `x-zenoh-key` on an event schema is its key.\n",
    "title": "dimos server",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/dimos/blueprints": {
      "get": {
        "description": "What `dimos list` prints: built-in blueprints (without demo-*), then external ones from installed packages. Cached for 60 s; `fresh` refills the cache first. No other side effects. `importable`, `import_error` and `missing_module` come from the discovery cache (null until the scan reaches the blueprint: GET /dimos/discovery).",
        "operationId": "get_dimos_blueprints",
        "parameters": [
          {
            "description": "check again now instead of answering from the cache",
            "in": "query",
            "name": "fresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "check again now instead of answering from the cache",
              "title": "Fresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintList"
                }
              }
            },
            "description": "`{ blueprints: [{ name, kind: \"builtin\"|\"external\", importable, import_error, missing_module }] }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the server refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Every blueprint dimos can run (name, builtin/external) and whether it imports",
        "tags": [
          "blueprints"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/discovery": {
      "get": {
        "description": "The scan starts when the server starts. A restart with the same checkout commit, dirty files and installed packages answers from the disk cache at once (no scan); a change is noticed within 30 s and rescanned (the old answer is served meanwhile, `stale: true`). `discovery` events follow it. No side effects.",
        "operationId": "get_dimos_discovery",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryStatus"
                }
              }
            },
            "description": "`DiscoveryStatus`: `{ state, reason, key, stale, blueprints_total, blueprints_done, importable, not_importable, modules_total, modules_done, current, errors, ... }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Where the discovery scan (every blueprint and module imported once, cached) is: progress, errors",
        "tags": [
          "discovery"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
    "/dimos/discovery/blueprints": {
      "get": {
        "description": "Answers from the cache (no import): every blueprint scanned so far, in registry order, with `suggested_extras` for one that's missing a package. No side effects.",
        "operationId": "get_dimos_discovery_blueprints",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveredBlueprints"
                }
              }
            },
            "description": "`{ stale, blueprints: [{ name, ref, robot, importable, import_error, missing_module, suggested_extras, modules: [{ name, class, module, streams: [{ name, type, direction, topic }] }] }] }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Every blueprint from the discovery cache: whether it imports (and why not), its robot, its modules and their streams with topics",
        "tags": [
          "discovery"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
    "/dimos/discovery/refresh": {
      "post": {
        "description": "Wakes the discovery loop: it recomputes the cache key and scans what's missing; with `full` it forgets the answer and imports everything again. Answers the status at once; `discovery` events follow the scan. The body is optional.",
        "operationId": "post_dimos_discovery_refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DiscoveryRefresh"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryStatus"
                }
              }
            },
            "description": "`DiscoveryStatus`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Check the checkout and packages again now, and rescan what changed (or everything, `full`)",
        "tags": [
          "discovery"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
    "/dimos/docs/custom-robot": {
      "get": {
        "description": "Found in the checkout's docs/ by file name and title (adding a custom/new robot, arm or platform); links point at the docs site (mkdocs.yml site_url), other repo files at GitHub. `html` is rendered here with markdown-it (raw HTML in the markdown is escaped), so a page can show it as is; `markdown` is there for a page that renders it itself. 404 when the checkout has no such page.",
        "operationId": "get_dimos_docs_custom_robot",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRobotDoc"
                }
              }
            },
            "description": "`{ title, markdown, html, source_path, url, others }`"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such thing (the message names it)"
          }
        },
        "summary": "dimos's guide to adding a robot of your own: markdown and HTML, with absolute links",
        "tags": [
          "docs"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
    "/dimos/docs/links": {
      "get": {
        "description": "Each link is the docs-site URL of the checkout's page by that name (docs/**/configuration.md, ...); null when there's no such page. No side effects.",
        "operationId": "get_dimos_docs_links",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocLinks"
                }
              }
            },
            "description": "`{ site, repo, configure_robot, custom_robot, blueprints, modules, installation, quickstart, cli }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Links into dimos's published docs: configuring a robot, adding one, blueprints, modules, install",
        "tags": [
          "docs"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/extras": {
      "get": {
        "description": "Extras from the checkout's pyproject.toml (else the installed dimos's metadata); installed packages are asked of the checkout's python in a child process (cached 10 s). `download_bytes` is an upper-bound hint from uv.lock. No side effects.",
        "operationId": "get_dimos_extras",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtrasList"
                }
              }
            },
            "description": "`{ mode, python, extras: [{ name, installed, applicable, requires, includes, missing, download_bytes }] }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "dimos's optional extras (`dimos[sim]`, ...): which are installed, what's missing, a download-size hint",
        "tags": [
          "extras"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
    "/dimos/extras/install": {
      "post": {
        "description": "Runs scripts/install.sh's command for them: `uv sync --locked --inexact --extra <x> ...` in a checkout (keeps every extra and group already installed), else `uv pip install --python <venv python> --torch-backend cpu|cu128 'dimos[<x>,...]==<version>'`. Never sudo. Answers the job at once: follow `<ns>/dimos/jobs/<job>` or GET /dimos/jobs/{job}/log; when it ends the extras cache is dropped and discovery rescans. 400 for an unknown extra; 409 while another install runs; 500 when uv isn't found.",
        "operationId": "post_dimos_extras_install",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtrasInstall"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStarted"
                }
              }
            },
            "description": "`{ job, command }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the server refuses (the message says which)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict with the current state (the message says what it is)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Install dimos extras, as a job (streamed output); discovery rescans when it ends",
        "tags": [
          "extras"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/jobs": {
      "get": {
        "description": "No side effects.",
        "operationId": "get_dimos_jobs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            },
            "description": "`{ jobs: [{ job, title, kind, done, ok, started_at, finished_at }] }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "The dimos server's jobs (extras installs): running, and finished in the last 30 minutes",
        "tags": [
          "jobs"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
...
    "/dimos/message-types": {
      "get": {
        "description": "Answers from the discovery cache. No side effects.",
        "operationId": "get_dimos_message_types",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageTypes"
                }
              }
            },
            "description": "`{ types: [{ type, publishers, subscribers }] }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Every message type on a module's stream, with the modules that publish and read it",
        "tags": [
          "discovery"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
  },
...
}